+++ src/main/java/im/enricods/ComicsStore/entities/Comic.java
@@ -1,6 +1,5 @@
 package im.enricods.ComicsStore.entities;
 
-import java.math.BigInteger;
 import java.util.Date;
 import java.util.Set;
 
@@ -31,7 +30,7 @@ public class Comic {
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id", nullable = false)
-    private BigInteger id;
+    private long id;
 
     @ManyToOne
     @JoinColumn(name = "collection")
